requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
google-generativeai==0.3.2
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
sentence-transformers==2.2.2
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0 